% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static.R
\name{pkg_missing_where}
\alias{pkg_missing_where}
\title{Locate calls to missing packages}
\usage{
pkg_missing_where(scan)
}
\arguments{
\item{scan}{Output from \code{pkg_scan()}.}
}
\value{
data.frame of package names and code files.
}
\description{
Locate calls to missing packages
}
