% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static.R
\name{pkg_scan}
\alias{pkg_scan}
\title{List all used packages in a directory and subdirectories}
\usage{
pkg_scan(paths = getwd())
}
\arguments{
\item{paths}{The path(s) to scan. Defaults to the current path.}
}
\value{
A list with
\itemize{
\item \code{called}: explicitly referenced packages (see \code{called$Packages})
\item \code{dependencies}: dependencies to \code{called}.
}
}
\description{
Looks for \code{library(package)}, \verb{package::function()} in .R and.Rmd files.
Looks for \code{Imports} etc. in DESCRIPTION files.
Identifies all dependencies too.

Does not scan .Rhistory and .md files

Technically, this function uses \code{renv::dependencies()} to scan files.
Then \verb{tools::package_dependencies()pkg, recursive = TRUE)} to find
dependencies.
}
