% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static.R
\name{pkg_unused}
\alias{pkg_unused}
\alias{pkg_missing}
\title{Installed packages that are not used}
\usage{
pkg_unused(scan, include_rstudio = TRUE)

pkg_missing(scan)
}
\arguments{
\item{scan}{Output from \code{pkg_scan()}.}

\item{include_rstudio}{Whether to also return packages used by RStudio (in \code{rstudioapi::getRStudioPackageDependencies()}).}
}
\value{
Character vector of package names.
}
\description{
Does not return packages in \code{getOption("defaultPackages")} nor the "compiler"
package.
}
\section{Functions}{
\itemize{
\item \code{pkg_missing}: Packages that are used but not installed
}}

