% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{register_pkg}
\alias{register_pkg}
\title{Register that this package was used}
\usage{
register_pkg(pkg)
}
\arguments{
\item{pkg}{Package name (character)}
}
\description{
This saves the usage to a global variable ".pkgusage"
}
\keyword{internal}
